+++ src/__tests__/GraphqlMockingService.test.ts
@@ -37,18 +37,47 @@ const schema = `
     }
 `;
 
+const subgraphSchema = `
+      type Query {
+        getRandomEmployee: Employee
+      }
+
+      type Employee {
+        name: String!
+      }
+
+      extend type Employee {
+        age: Int!
+      }
+
+      extend type Company @key(fields: "companyId") {
+        companyId: Int! @external
+        employees: [Employee]
+      }
+    `;
+
 describe('GraphqlMockingService', () => {
   const port = 3001;
+  const subgraphPort = 3002;
   let mockingService;
+  let subgraphMockingService;
   const sequenceId = 'test-sequence-id';
   beforeAll(async () => {
     mockingService = new GraphqlMockingService({port});
     await mockingService.start();
     await mockingService.registerSchema(schema);
+
+    subgraphMockingService = new GraphqlMockingService({
+      port: subgraphPort,
+      subgraph: true,
+    });
+    await subgraphMockingService.start();
+    await subgraphMockingService.registerSchema(subgraphSchema);
   });
 
   afterAll(async () => {
     await mockingService.stop();
+    await subgraphMockingService.stop();
   });
 
   it('should allow operation seed registration', async () => {
@@ -484,5 +513,54 @@ describe('GraphqlMockingService', () => {
 
     expect(secondOperationResult).toEqual(secondMock);
     expect(thirdOperationResult).toEqual({data: thirdMock});
+  });
+
+  it('should support subgraph schemas', async () => {
+    const mockingContext = subgraphMockingService.createContext();
+    const operationName = 'getEmployee';
+    await mockingContext.operation(
+      operationName,
+      {
+        data: {
+          getRandomEmployee: {
+            name: 'John',
+          },
+        },
+      },
+      {}
+    );
+
+    const operationResult = await fetch(
+      `http://localhost:${subgraphPort}/graphql`,
+      {
+        method: 'post',
+        body: JSON.stringify({
+          operationName,
+          query: 'query getEmployee { getRandomEmployee { name } }',
+          variables: {},
+        }),
+        headers: {
+          'Content-Type': 'application/json',
+          'mocking-sequence-id': mockingContext.sequenceId,
+        },
+      }
+    ).then((res) => res.json());
+
+    expect(operationResult).toEqual({
+      data: {
+        getRandomEmployee: {
+          name: 'John',
+        },
+      },
+    });
+
+    await subgraphMockingService.stop();
+  });
+
+  it('should allow creating contexts with a shared sequenceId', function () {
+    const contextA = mockingService.createContext();
+    const contextB = subgraphMockingService.createContext(contextA.sequenceId);
+
+    expect(contextA.sequenceId === contextB.sequenceId);
   });
 });